Using truetype font instead of images for keyboard keys
To make internationalization I decided to throw out all images for the
keys and insted use a truetype font.
The font used is Arial with some changes:
1. Some characters removed to save space
2. New characters created for the keys HIDE, UPPERCASE, LOWERCASE,
DELETE and ENTER (the last not yet implemented in code)

By creating two text fields slightly offset fom each other the
"shadow-effect" is still preserved in the keys.